refactor(reservation): extract pricing utils from store
- Create frontend/src/utils/pricing.ts with daysBetween, computePriceBreakdown, formatPrice
- Replace function bodies in reservationStore with re-exports from new utils
- Zero behavior change, same function signatures, no consumer broken

PR1/4 of refactor/reservation-flow chain
